+++ cmd/partitiontest_linter/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/algorand/go-algorand/cmd/testpartitioning_linter/analyzer"
+	"github.com/algorand/go-algorand/cmd/partitiontest_linter/analyzer"
 	"golang.org/x/tools/go/analysis/singlechecker"
 )
